Added HttpClient provider in app.config.ts and reorganized the structure of the application
