+++ front/src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { AppRoutes } from './app.routes';
+import { AppRoutes } from './core/app.routes';
 import { AppComponent } from './app.component';
 import { HomeComponent } from './features/home/home.component';
