+++ front/src/app/core/app.config.ts
@@ -1,0 +1,11 @@
+import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
+import { provideRouter } from '@angular/router';
+
+import { routes } from './app.routes';
+import {provideHttpClient} from "@angular/common/http";
+import {provideAnimations} from "@angular/platform-browser/animations";
+
+export const appConfig: ApplicationConfig = {
+  providers: [provideZoneChangeDetection({ eventCoalescing: true }), provideRouter(routes), provideHttpClient(),
+    provideAnimations()]
+};
